backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
PyPDF2==3.0.1
faiss-cpu>=1.7.0
numpy==1.26.4
python-dotenv==1.0.0
scikit-learn==1.3.2
Pillow==10.1.0
aiofiles==23.2.1